+++ src/main/java/com/robertx22/age_of_exile/database/data/currency/CrystalOfPurificationItem.java
@@ -4,7 +4,6 @@ import com.robertx22.age_of_exile.database.data.currency.base.CurrencyItem;
 import com.robertx22.age_of_exile.database.data.currency.base.ICurrencyItemEffect;
 import com.robertx22.age_of_exile.database.data.currency.base.IShapedRecipe;
 import com.robertx22.age_of_exile.database.data.currency.loc_reqs.BaseLocRequirement;
-import com.robertx22.age_of_exile.database.data.currency.loc_reqs.GearEnumLocReq;
 import com.robertx22.age_of_exile.database.data.currency.loc_reqs.SimpleGearLocReq;
 import com.robertx22.age_of_exile.database.data.currency.loc_reqs.item_types.GearReq;
 import com.robertx22.age_of_exile.mmorpg.ModRegistry;
@@ -14,47 +13,38 @@ import com.robertx22.age_of_exile.uncommon.datasaving.Gear;
 import com.robertx22.age_of_exile.uncommon.interfaces.data_items.IRarity;
 import net.minecraft.data.server.recipe.ShapedRecipeJsonFactory;
 import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
 
 import java.util.Arrays;
 import java.util.List;
 
-public class OrbOfDisorder extends CurrencyItem implements ICurrencyItemEffect, IShapedRecipe {
+public class CrystalOfPurificationItem extends CurrencyItem implements ICurrencyItemEffect, IShapedRecipe {
     @Override
     public String GUID() {
-        return "currency/randomize_affixes";
+        return "currency/crystal_of_purification";
     }
 
-    private static final String name = Ref.MODID + ":currency/randomize_affixes";
-
-    public OrbOfDisorder() {
-
-        super(name);
-
-    }
+    public static final String ID = Ref.MODID + ":currency/crystal_of_purification";
 
     @Override
     public int getWeight() {
-        return 250;
+        return 500;
+    }
+
+    public CrystalOfPurificationItem() {
+        super(ID);
     }
 
     @Override
     public ItemStack ModifyItem(ItemStack stack, ItemStack Currency) {
         GearItemData gear = Gear.Load(stack);
-        gear.affixes.randomize(gear);
+        gear.sealed = false;
         Gear.Save(stack, gear);
-
         return stack;
     }
 
     @Override
     public List<BaseLocRequirement> requirements() {
-        return Arrays.asList(GearReq.INSTANCE, SimpleGearLocReq.IS_NOT_UNIQUE, GearEnumLocReq.AFFIXES);
-    }
-
-    @Override
-    public float getInstability() {
-        return 15;
+        return Arrays.asList(GearReq.INSTANCE, SimpleGearLocReq.IS_SEALED);
     }
 
     @Override
@@ -63,32 +53,35 @@ public class OrbOfDisorder extends CurrencyItem implements ICurrencyItemEffect, 
     }
 
     @Override
+    public float getInstability() {
+        return 0;
+    }
+
+    @Override
     public int getRarityRank() {
-        return IRarity.Common;
+        return IRarity.Magical;
     }
 
     @Override
     public String locNameForLangFile() {
-        return nameColor + "Orb Of Disorder";
+        return nameColor + "Crystal of Purification";
     }
 
     @Override
     public String locDescForLangFile() {
-        return "Re-rolls All affixes";
+        return "Unseals the item.";
     }
 
     @Override
     public ShapedRecipeJsonFactory getRecipe() {
-        return ShapedRecipeJsonFactory.create(ModRegistry.CURRENCIES.ORB_OF_DISORDER, 1)
-            .input('#', ModRegistry.MISC_ITEMS.INFUSED_IRON)
-            .input('t', ModRegistry.CURRENCIES.ORB_OF_TRANSMUTATION)
-            .input('v', Items.GOLD_NUGGET)
+        return shaped(ModRegistry.CURRENCIES.CRYSTAL_OF_PURIFICATION)
+            .input('t', ModRegistry.CURRENCIES.CRYSTAL_OF_TRUTH)
+            .input('v', ModRegistry.GEAR_MATERIALS.ARCANA)
             .input('o', ModRegistry.MISC_ITEMS.MAGIC_ESSENCE)
-            .pattern("o#o")
-            .pattern("oto")
-            .pattern("vvv")
+            .pattern("ovo")
+            .pattern("vtv")
+            .pattern("ovo")
             .criterion("player_level", trigger());
-
     }
 
 }