+++ src/main/java/com/robertx22/age_of_exile/saveclasses/unit/GearData.java
@@ -69,7 +69,7 @@ public class GearData {
         BaseGearType type = gear.GetBaseGearType();
 
         if (type.isMeleeWeapon()) {
-            return slot == EquipmentSlot.MAINHAND || slot == EquipmentSlot.OFFHAND;
+            return slot == EquipmentSlot.MAINHAND;
         }
         if (type.isWeapon()) {
             return slot == EquipmentSlot.MAINHAND; // ranged weapon